fix: 系统初始安装时脚本错误导致安装失败
系统预安装应用初始没有这个进程导致killall脚本失败，所以killall之前判断一下进场是否存在。

Log: 解决rpm版预安装失败

+++ rpm/dde-printer.spec
@@ -52,8 +52,10 @@ popd
 %install
 %make_install -C build INSTALL_ROOT="%buildroot"
 
-%post
-killall  -USR1  dde-printer-helper
+%posttrans
+if [ -n "$(ps -ef | grep dde-printer | grep -v grep)" ];then 
+    killall -USR1 dde-printer-helper
+fi
 
 %files
 %doc README.md
